error handling to api delete scenario stubs

+++ api.go
@@ -41,21 +41,24 @@ func (c *Client) getScenarioStubs(scenario string) ([]byte, error) {
 func (c *Client) deleteScenarioStubs(p apiParams) ([]byte, error) {
 	var s params
 	// adding path
-	path := "/stubo/api/v2/scenarios/objects/" + p.name + "/stubs"
-	s.path = path
-	// creating MAP for headers
-	headers := make(map[string]string)
-	if p.force != "" {
-		headers["force"] = p.force
+	if p.name != "" {
+		path := "/stubo/api/v2/scenarios/objects/" + p.name + "/stubs"
+		s.path = path
+		// creating MAP for headers
+		headers := make(map[string]string)
+		if p.force != "" {
+			headers["force"] = p.force
+		}
+		if p.targetHost != "" {
+			headers["target_host"] = p.targetHost
+		}
+		s.headers = headers
+		s.method = "DELETE"
+		fmt.Println(s.headers)
+		// calling delete
+		return c.makeRequest(s)
 	}
-	if p.targetHost != "" {
-		headers["target_host"] = p.targetHost
-	}
-	s.headers = headers
-	s.method = "DELETE"
-	fmt.Println(s.headers)
-	// calling delete
-	return c.makeRequest(s)
+	return []byte(""), ErrMissingParams
 }
 
 // getDelayPolicy gets specified delay-policy
